+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.0.11
+
+* Better README and docs
+  * Everything now refers to the top level callback as the *callbackDispatcher* 
+  * Some typo's 
+  * Link to [Medium blog](https://medium.com/vrt-digital-studio/flutter-workmanager-81e0cfbd6f6e) post
+* Adds unit tests to the project since the many issues with enum parsing
+  * Fixes an issue with the parsing of the `NetworkType`
+* Project restructure to a `src` folder 
+
 # 0.0.10
 
 * I should test a little better before publishing to pub.dev.